refactor: 波形接続を `tts_pipeline` へ移動
