+++ test/tts_pipeline/test_connect_base64_waves.py
@@ -8,7 +8,7 @@ import soundfile
 from numpy.typing import NDArray
 from soxr import resample
 
-from voicevox_engine.utility.connect_base64_waves import (
+from voicevox_engine.tts_pipeline.connect_base64_waves import (
     ConnectBase64WavesException,
     connect_base64_waves,
 )
